Bug fix on deleting mysql databases - related to transitioning to new fast tables
git-svn-id: http://192.168.1.10/repos/toto@1718 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/DatabaseServer/src/com/azquo/memorydb/dao/FastDAO.java
@@ -55,12 +55,13 @@ public abstract class FastDAO {
     }
 
     public void clearTable(final String databaseName){
-        jdbcTemplateUtils.update("delete from `" + databaseName + "`.`" + getTableName() + "`", JsonRecordDAO.EMPTY_PARAMETERS_MAP);
-
+        if (checkFastTableExists(databaseName)){
+            jdbcTemplateUtils.update("delete from `" + databaseName + "`.`" + getTableName() + "`", JsonRecordDAO.EMPTY_PARAMETERS_MAP);
+        }
     }
 
-    public boolean checkFastTableExists(final AzquoMemoryDB azquoMemoryDB){
-        final List<Map<String, Object>> maps = jdbcTemplateUtils.queryForList("show tables from  `" + azquoMemoryDB.getMySQLName() + "` like 'fast_value' ", JsonRecordDAO.EMPTY_PARAMETERS_MAP);
+    public boolean checkFastTableExists(final String databaseName){
+        final List<Map<String, Object>> maps = jdbcTemplateUtils.queryForList("show tables from  `" + databaseName + "` like 'fast_value' ", JsonRecordDAO.EMPTY_PARAMETERS_MAP);
         return !maps.isEmpty();
     }
